Биллинг: Amount_t теперь Long.
До этого - был Double со всеми его глюками.

Теперь цены хранятся в копейках, центах и т.д.
Это решает проблему долей копеек, которые естественным образом
появляются в floatingPoint-типах после арифметических операций,
но неприемлемы в биллинге, т.к. можно списать лишнюю копейку и даже
уйти в минус там, где это по человеческой логике было невозможно.

Допиливание кода под новый целочисленный тип: форматирование работает.
Из проблем: форму ввода цен и оплату - пока надо протестировать.
Sys-формы пока принимают деньги только в копейках, и пока не ясно, как
это красиво разрулить, т.к. form-mapping'и валюты и amount'а цены находятся
на разных уровнях кода форм.

Перед обновлением необходимо накатить sql-код миграции: 16.sql.

+++ src1/client/lk/nodes/form/src/main/scala/io/suggest/lk/nodes/form/a/pop/EditTfDailyAh.scala
@@ -36,7 +36,7 @@ class EditTfDailyAh[M](
           .orElse( tf.clauses.values.headOption )
       }
     val priceAmount = priceOpt
-      .fold[Amount_t](1.0) { _.amount }
+      .fold[Amount_t](1) { _.amount }
     val mprice = priceOpt.getOrElse {
       MPrice(priceAmount, nodeTfOpt.get.currency )
     }
@@ -56,8 +56,12 @@ class EditTfDailyAh[M](
       val v0 = value
       val v2 = for (v <- v0) yield {
         val (mode2, isValid) = try {
-          val amount2 = m.amount.trim.replace(',' , '.').toDouble
-          val mode22 = v.mode.manualOpt.fold( ManualTf(amount2) )( _.withAmount(amount2) )
+          val amount2 = m.amount
+            .trim
+            .replace(',' , '.')
+            .toLong
+          val mode22 = v.mode.manualOpt
+            .fold( ManualTf(amount2) )( _.withAmount(amount2) )
           mode22 -> true
         } catch {
           case _: Throwable =>
